Resultado Final

+++ exercicio_1.php
@@ -1,6 +1,6 @@
 <?php
 
 print "Qual a medida em metros:";
-$metros=fgets(STDIN);
+$metros=(float)fgets(STDIN);
 $milimetros=$metros*1000;
 print "São $milimetros mm";
